+++ .github/workflows/publish-release.yml
@@ -117,4 +117,4 @@ jobs:
         uses: actions/upload-artifact@v2
         with:
           name: Maven publishing artifacts for ${{ matrix.branch }}
-          path: publish/offers-list/
+          path: publish/offers-hud/
