+++ src/main/java/net/naari3/offershud/mixin/CloseScreen.java
@@ -1,17 +1,17 @@
-package net.naari3.offerslist.mixin;
+package net.naari3.offershud.mixin;
 
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.At;
 
-import net.naari3.offerslist.OffersList;
+import net.naari3.offershud.OffersHUD;
 import net.minecraft.client.network.ClientPlayerEntity;
 
 @Mixin(ClientPlayerEntity.class)
 abstract class CloseScreen {
     @Inject(at = @At("HEAD"), method = "closeHandledScreen")
     public void closeHandledScreen(CallbackInfo ci) {
-        OffersList.setOpenWindow(false);
+        OffersHUD.setOpenWindow(false);
     }
 }
