+++ src/main/resources/assets/offershud/icon.png
